Ensure we adjust modtime on unchanged files.
Currently we don't overwrite files if the hash is the same as an existing
file. This leads to issues when we clean the dir since we may delete
the current file as it's old, but we should rather keep it. Change
the behaviour to adjust  the date on unchanged files to prevent this.